LPS-42120 Rename
